layout algorithm draw out
