+++ library/src/main/java/com/gyso/treeview/model/TreeModel.java
@@ -1,8 +1,10 @@
 package com.gyso.treeview.model;
 
+import android.app.ActionBar;
 import android.util.SparseArray;
 import android.util.SparseIntArray;
 
+import com.gyso.treeview.algorithm.table.Table;
 import com.gyso.treeview.util.TreeViewLog;
 
 import java.io.Serializable;
@@ -110,201 +112,6 @@ public class TreeModel<T> implements Serializable {
         return maxChildNode;
     }
 
-    public void calculateTreeNodesDeep(){
-        //calculateTreeNodesDeepLoose();
-        calculateTreeNodesDeepCompact();
-    }
-    /**
-     * calculate the deep of all tree nodes  by loose
-     * deep has sort 0--->n
-     */
-    public void calculateTreeNodesDeepLoose(){
-        TreeViewLog.e(TAG,"calculateTreeNodesDeep start");
-        Stack<NodeModel<T>> stack = new Stack<>();
-        NodeModel<T> rootNode = getRootNode();
-        stack.add(rootNode);
-        while (!stack.isEmpty()) {
-            NodeModel<T> cur = stack.pop();
-            NodeModel<T> parentNode = cur.getParentNode();
-            cur.deep = 0;
-            //if has peer, peer.deep+peer.leafCount+1/*peer self*/
-            //if has no peer, parentNode.deep;
-            if(parentNode!=null){
-                int indexOfCur = parentNode.childNodes.indexOf(cur);
-                if(indexOfCur>0){
-                    NodeModel<T> prePeer = parentNode.childNodes.get(indexOfCur - 1);
-                    cur.deep += prePeer.deep;
-                    cur.deep += (prePeer.leafCount ==0?1:prePeer.leafCount);
-                }else{
-                    cur.deep = parentNode.deep;
-                }
-            }
-            TreeViewLog.e(TAG,"calculateTreeNodesDeep--->"+cur.toString());
-            LinkedList<NodeModel<T>> childNodes = cur.getChildNodes();
-            for (int i = childNodes.size()-1; i >=0; i--) {
-                stack.add(childNodes.get(i));
-            }
-        }
-        TreeViewLog.e(TAG,"calculateTreeNodesDeep end");
-    }
-
-    /**
-     * calculate the deep of all tree nodes  by compact way
-     * deep has sort 0--->n
-     */
-    public void calculateTreeNodesDeepCompact(){
-        TreeViewLog.e(TAG,"calculateTreeNodesDeep start");
-        Deque<NodeModel<T>> deque = new ArrayDeque<>();
-        NodeModel<T> rootNode = getRootNode();
-        deque.add(rootNode);
-        SparseIntArray deepSum = new SparseIntArray();
-        //calculate base deep
-        while (!deque.isEmpty()) {
-            NodeModel<T> cur = deque.poll();
-            cur.deep = deepSum.get(cur.floor,0);
-            recordDeep(cur);
-            deepSum.put(cur.floor,cur.deep+1);
-            LinkedList<NodeModel<T>> childNodes = cur.getChildNodes();
-            if (childNodes.size() > 0) {
-                deque.addAll(childNodes);
-            }
-        }
-        compactTable();
-        TreeViewLog.e(TAG,"calculateTreeNodesDeepCompact end");
-    }
-
-    private void recordDeep(NodeModel<?> node) {
-        if(node==null){
-            return;
-        }
-        maxDeep = Math.max(node.deep, maxDeep);
-        minDeep = Math.min(node.deep, minDeep);
-    }
-
-    public int getMinDeep(){
-        return minDeep;
-    }
-
-    public int getMaxDeep(){
-        return maxDeep;
-    }
-
-
-    private void compactTable(){
-        //calculate final deep
-        Deque<NodeModel<T>> deque = new ArrayDeque<>();
-        LinkedList<NodeModel<T>> hasDealNode  = new LinkedList<>();
-        deque.add(rootNode);
-        SparseIntArray deepSum = new SparseIntArray();
-        while (!deque.isEmpty()) {
-            NodeModel<T> cur = deque.poll();
-            if(hasDealNode.contains(cur)){
-                continue;
-            }else{
-                hasDealNode.add(cur);
-            }
-            int peerMidDeep;
-            int parentDeep;
-
-            //deal parent and me
-            NodeModel<T> parentNode = cur.getParentNode();
-            if(parentNode!=null){
-                LinkedList<NodeModel<T>> peers = parentNode.getChildNodes();
-                if(peers !=null && !peers.isEmpty()){
-                    int sum=0;
-                    for(NodeModel<T> peer: peers){
-                        sum +=peer.deep;
-                    }
-                    peerMidDeep=sum/peers.size();
-                    parentDeep = parentNode.deep;
-                    if(parentDeep!=peerMidDeep){
-                        if (parentDeep<peerMidDeep){
-                            //parent all left move to peers' mid
-                            final int d = peerMidDeep-parentDeep;
-                            fromRootToMyUpRight(cur, next -> {
-                                next.deep +=d;
-                                recordDeep(next);
-                                //next's parent fit center
-                                NodeModel<?> np = next.getParentNode();
-                                int nSum=0;
-                                if(np!=null){
-                                    LinkedList<? extends NodeModel<?>> npChildNodes = np.getChildNodes();
-                                    for(NodeModel<?> nPeer:npChildNodes ){
-                                        nSum +=nPeer.deep;
-                                    }
-                                    np.deep = nSum/npChildNodes.size();
-                                    recordDeep(np);
-                                }
-                            });
-                        }else{
-                            final int d = parentDeep - peerMidDeep;
-                            //peers' mid move to parent
-                            LinkedList<NodeModel<T>> nodeModels = arrayByFloor.get(cur.floor);
-                            int md = cur.deep;
-                            for (NodeModel<T> afterMe:nodeModels) {
-                                if(afterMe.deep >= md){
-                                    afterMe.deep += d;
-                                    recordDeep(afterMe);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            LinkedList<NodeModel<T>> childNodes = cur.getChildNodes();
-            maxDeep = Math.max(cur.deep, maxDeep);
-            deepSum.put(cur.floor,cur.deep+1);
-            if (childNodes.size() > 0) {
-                deque.addAll(childNodes);
-            }
-        }
-    }
-
-    private void fromRootToMyUpRight(final NodeModel<T> node, ITraversal<NodeModel<?>> traversal){
-        //mark line
-        NodeModel<?> markNode = node;
-        SparseIntArray modelSparseArray = new SparseIntArray();
-        while (markNode!=null){
-            modelSparseArray.put(markNode.floor,markNode.deep);
-            markNode = markNode.parentNode;
-        }
-        Deque<NodeModel<T>> deque = new ArrayDeque<>();
-        NodeModel<T> tmpNode = getRootNode();
-        deque.add(tmpNode);
-        while (!deque.isEmpty()) {
-            tmpNode = deque.poll();
-            int markDeep= modelSparseArray.get(tmpNode.floor);
-            if(markDeep>tmpNode.deep){
-                continue;
-            }
-            if(tmpNode.floor>=node.floor){
-                break;
-            }
-            if (traversal != null) {
-                traversal.next(tmpNode);
-            }
-            if(tmpNode==null){
-                continue;
-            }
-            LinkedList<NodeModel<T>> childNodes = tmpNode.getChildNodes();
-            if (childNodes.size() > 0) {
-                deque.addAll(childNodes);
-            }
-        }
-        if (traversal != null) {
-            traversal.finish();
-        }
-    }
-
-    private void moveDeepForRightAndUp(){
-
-    }
-
-    private void moveDeepForRightAndDown(){
-
-    }
-
-
     /**
      *child nodes will ergodic in the last
      * 广度遍历
@@ -336,6 +143,10 @@ public class TreeModel<T> implements Serializable {
             iTraversal.finish();
             this.finishTraversal = false;
         }
+    }
+
+    public SparseArray<LinkedList<NodeModel<T>>> getArrayByFloor() {
+        return arrayByFloor;
     }
 
     /**
@@ -396,5 +207,21 @@ public class TreeModel<T> implements Serializable {
             iTraversal.finish();
             this.finishTraversal = false;
         }
+    }
+
+    public int getMaxDeep() {
+        return maxDeep;
+    }
+
+    public void setMaxDeep(int maxDeep) {
+        this.maxDeep = maxDeep;
+    }
+
+    public int getMinDeep() {
+        return minDeep;
+    }
+
+    public void setMinDeep(int minDeep) {
+        this.minDeep = minDeep;
     }
 }
